wip: consolidate path
