+++ src/routing/path/create-path-catalog.ts
